pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-interface"
version = "0.1.0"
description = "A multi-provider LLM web interface with tool use, conversation management, and streaming chat."
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.30.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "websockets>=13.0",
    "feedparser>=6.0.0",
    "python-multipart>=0.0.12",
    "pymupdf>=1.25.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8.0",
]


[tool.setuptools.packages.find]
where = ["."]
include = ["llm_interface*"]

[project.scripts]
llm-interface = "llm_interface.server:main"
llm-interface-briefing = "llm_interface.briefing_assembly:cli_main"
llm-interface-email = "llm_interface.email_ingestion:cli_main"
